+++ src/material/checkbox/testing/BUILD.bazel
@@ -14,19 +14,6 @@ ts_library(
     ],
 )
 
-ng_test_library(
-    name = "harness_tests_lib",
-    srcs = ["shared.spec.ts"],
-    deps = [
-        ":testing",
-        "//src/cdk/testing",
-        "//src/cdk/testing/testbed",
-        "//src/material/checkbox",
-        "@npm//@angular/forms",
-        "@npm//@angular/platform-browser",
-    ],
-)
-
 filegroup(
     name = "source-files",
     srcs = glob(["**/*.ts"]),
@@ -34,14 +21,14 @@ filegroup(
 
 ng_test_library(
     name = "unit_tests_lib",
-    srcs = glob(
-        ["**/*.spec.ts"],
-        exclude = ["shared.spec.ts"],
-    ),
+    srcs = glob(["**/*.spec.ts"]),
     deps = [
-        ":harness_tests_lib",
         ":testing",
+        "//src/cdk/testing",
+        "//src/cdk/testing/testbed",
         "//src/material/checkbox",
+        "@npm//@angular/forms",
+        "@npm//@angular/platform-browser",
     ],
 )
